@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { FormControl, FormGroup } from "@angular/forms";
+import { UntypedFormControl, UntypedFormGroup } from "@angular/forms";
 import { map, Observable, startWith } from "rxjs";
 import { ExchangeService } from "@app/pages/exchange/services/exchange.service";
 import { ICurrentExchange, IDailyExchange } from "@app/pages/exchange/interfaces/exchange.interface";
@@ -22,7 +22,7 @@ import { MatSnackBar } from "@angular/material/snack-bar";
 export class ExchangeComponent {
   public currencyOptions = ["AED", "AFN", "ALL", "AMD", "ANG", "AOA", "ARS", "AUD", "AWG", "AZN", "BAM", "BBD", "BDT", "BGN", "BHD", "BIF", "BMD", "BND", "BOB", "BOV", "BRL", "BSD", "BTN", "BWP", "BYR", "BZD", "CAD", "CDF", "CHE", "CHF", "CHW", "CLF", "CLP", "CNY", "COP", "COU", "CRC", "CUC", "CUP", "CVE", "CZK", "DJF", "DKK", "DOP", "DZD", "EGP", "ERN", "ETB", "EUR", "FJD", "FKP", "GBP", "GEL", "GHS", "GIP", "GMD", "GNF", "GTQ", "GYD", "HKD", "HNL", "HRK", "HTG", "HUF", "IDR", "ILS", "INR", "IQD", "IRR", "ISK", "JMD", "JOD", "JPY", "KES", "KGS", "KHR", "KMF", "KPW", "KRW", "KWD", "KYD", "KZT", "LAK", "LBP", "LKR", "LRD", "LSL", "LTL", "LVL", "LYD", "MAD", "MDL", "MGA", "MKD", "MMK", "MNT", "MOP", "MRO", "MUR", "MVR", "MWK", "MXN", "MXV", "MYR", "MZN", "NAD", "NGN", "NIO", "NOK", "NPR", "NZD", "OMR", "PAB", "PEN", "PGK", "PHP", "PKR", "PLN", "PYG", "QAR", "RON", "RSD", "RUB", "RWF", "SAR", "SBD", "SCR", "SDG", "SEK", "SGD", "SHP", "SLL", "SOS", "SRD", "SSP", "STD", "SYP", "SZL", "THB", "TJS", "TMT", "TND", "TOP", "TRY", "TTD", "TWD", "TZS", "UAH", "UGX", "USD", "USN", "USS", "UYI", "UYU", "UZS", "VEF", "VND", "VUV", "WST", "XAF", "XAG", "XAU", "XBA", "XBB", "XBC", "XBD", "XCD", "XDR", "XFU", "XOF", "XPD", "XPF", "XPT", "XTS", "XXX", "YER", "ZAR", "ZMW"];
   public filteredOptions: Observable<string[]> | undefined;
-  public form: FormGroup;
+  public form: UntypedFormGroup;
   public openList: boolean = false;
   public currentData: ICurrentExchange | undefined;
   public dailyItems: IDailyExchange[] = [];
@@ -32,8 +32,8 @@ export class ExchangeComponent {
     private exchangeService: ExchangeService,
     private snackBarService: MatSnackBar
   ) {
-    this.form = new FormGroup({
-      currency: new FormControl('', []),
+    this.form = new UntypedFormGroup({
+      currency: new UntypedFormControl('', []),
     });
 
     this.filteredOptions = this.form.get('currency')?.valueChanges.pipe(
